move implementation of services command to services package
